requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.2
django-js-asset==2.0.0
django-mptt==0.14.0
gunicorn==20.1.0
sqlparse==0.4.3
tzdata==2023.2
whitenoise==6.5.0
